+++ .gitignore
@@ -122,3 +122,5 @@ debug
 
 # pytest cahche
 .pytest_cache
+/docs
+/ot/mytest
